refactor(data): Refactor UserRemoteDataSource and introduce request models
